pyproject.toml
[project]
name = "multi-system-agent-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "chromadb>=1.5.5",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.2",
    "ipykernel>=7.2.0",
    "langchain>=1.2.13",
    "langgraph>=1.1.3",
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "openai>=2.1.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "tavily-python>=0.7.23",
    "tiktoken>=0.12.0",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "autopep8>=2.3.2",
    "isort>=8.0.1",
    "pytest>=8.4.2",
    "ruff>=0.15.7",
]

[tool.ruff]
line-length = 120
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C", "PL", "PT"]
ignore = ["PLR0913", "PLR0912", "PLR0915", "PLR0917", "PLR0911", "PLR0914", "PLR0916", "PLR0910", "PLR0918", "PLR0919", "PLR0920", "PLR0921", "PLR0922", "PLR0923", "PLR0924", "PLR0925", "PLR0926", "PLR0927", "PLR0928", "PLR0929", "PLR0930", "PLR0931", "PLR0932", "PLR0933", "PLR0934", "PLR0935", "PLR0936", "PLR0937", "PLR0938", "PLR0939", "PLR0940", "PLR0941", "PLR0942", "PLR0943", "PLR0944", "PLR0945", "PLR0946", "PLR0947", "PLR0948", "PLR0949", "PLR0950", "PLR0951", "PLR0952", "PLR0953", "PLR0954", "PLR0955", "PLR0956", "PLR0957", "PLR0958", "PLR0959", "PLR0960", "PLR0961", "PLR0962", "PLR0963", "PLR0964", "PLR0965", "PLR0966", "PLR0967", "PLR0968", "PLR0969", "PLR0970", "PLR0971", "PLR0972", "PLR0973", "PLR0974", "PLR0975", "PLR0976", "PLR0977", "PLR0978", "PLR0979", "PLR0980", "PLR0981", "PLR0982", "PLR0983", "PLR0984", "PLR0985", "PLR0986", "PLR0987", "PLR0988", "PLR0989", "PLR0990", "PLR0991", "PLR0992", "PLR0993", "PLR0994", "PLR0995", "PLR0996", "PLR0997", "PLR0998", "PLR0999", "PLR1000", "PLR1001", "PLR1002", "PLR1003", "PLR1004", "PLR1005", "PLR1006", "PLR1007", "PLR1008", "PLR1009", "PLR1010", "PLR1011", "PLR1012", "PLR1013", "PLR1014", "PLR1015", "PLR1016", "PLR1017", "PLR1018", "PLR1019", "PLR1020", "PLR1021", "PLR1022", "PLR1023", "PLR1024", "PLR1025", "PLR1026", "PLR1027", "PLR1028", "PLR1029", "PLR1030", "PLR1031", "PLR1032", "PLR1033", "PLR1034", "PLR1035", "PLR1036", "PLR1037", "PLR1038", "PLR1039", "PLR1040", "PLR1041", "PLR1042", "PLR1043", "PLR1044", "PLR1045", "PLR1046", "PLR1047", "PLR1048", "PLR1049", "PLR1050", "PLR1051", "PLR1052", "PLR1053", "PLR1054", "PLR1055", "PLR1056", "PLR1057", "PLR1058", "PLR1059", "PLR1060", "PLR1061", "PLR1062", "PLR1063", "PLR1064", "PLR1065", "PLR1066", "PLR1067", "PLR1068", "PLR1069", "PLR1070", "PLR1071", "PLR1072", "PLR1073", "PLR1074", "PLR1075", "PLR1076", "PLR1077", "PLR1078", "PLR1079", "PLR1080", "PLR1081", "PLR1082", "PLR1083", "PLR1084", "PLR1085", "PLR1086", "PLR1087", "PLR1088", "PLR1089", "PLR1090", "PLR1091", "PLR1092", "PLR1093", "PLR1094", "PLR1095", "PLR1096", "PLR1097", "PLR1098", "PLR1099", "PLR1100", "PLR1101", "PLR1102", "PLR1103", "PLR1104", "PLR1105", "PLR1106", "PLR1107", "PLR1108", "PLR1109", "PLR1110", "PLR1111", "PLR1112", "PLR1113", "PLR1114", "PLR1115", "PLR1116", "PLR1117", "PLR1118", "PLR1119", "PLR1120", "PLR1121", "PLR1122", "PLR1123", "PLR1124", "PLR1125", "PLR1126", "PLR1127", "PLR1128", "PLR1129", "PLR1130", "PLR1131", "PLR1132", "PLR1133", "PLR1134", "PLR1135", "PLR1136", "PLR1137", "PLR1138", "PLR1139", "PLR1140", "PLR1141", "PLR1142", "PLR1143", "PLR1144", "PLR114"]
